models/requirements.txt
streamlit==1.25.0
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.8.0.76
scikit-learn==1.3.2
pillow==9.5.0
pandas==1.5.3
altair==5.1.2
pydeck==0.8.1b0
openpyxl==3.1.2
